Move some code to more logical locations; all tests pass
