+++ gslides_api/element/element.py
@@ -18,7 +18,11 @@ from gslides_api.element.base import PageElementBase, ElementKind
 from gslides_api.client import api_client
 from gslides_api.element.shape import ShapeElement
 from gslides_api.utils import dict_to_dot_separated_field_list, image_url_is_valid
-from gslides_api.request.request import UpdateImagePropertiesRequest, ReplaceImageRequest
+from gslides_api.request.request import (
+    UpdateImagePropertiesRequest,
+    ReplaceImageRequest,
+    GslidesAPIRequest,
+)
 
 
 def element_discriminator(v: Any) -> str:
@@ -93,10 +97,10 @@ class TableElement(PageElementBase):
             }
         ]
 
-    def element_to_update_request(self, element_id: str) -> List[Dict[str, Any]]:
+    def element_to_update_request(self, element_id: str) -> List[GslidesAPIRequest]:
         """Convert a TableElement to an update request for the Google Slides API."""
         # Tables don't have specific properties to update beyond base properties
-        return self.get_base_update_requests(element_id)
+        return self.alt_text_update_request(element_id)
 
 
 class ImageElement(PageElementBase):
@@ -124,9 +128,9 @@ class ImageElement(PageElementBase):
             }
         ]
 
-    def element_to_update_request(self, element_id: str) -> List[Dict[str, Any]]:
+    def element_to_update_request(self, element_id: str) -> List[GslidesAPIRequest]:
         """Convert an ImageElement to an update request for the Google Slides API."""
-        requests = self.get_base_update_requests(element_id)
+        requests = self.alt_text_update_request(element_id)
 
         if hasattr(self.image, "imageProperties") and self.image.imageProperties is not None:
             image_properties = self.image.imageProperties.to_api_format()
@@ -134,9 +138,9 @@ class ImageElement(PageElementBase):
             request = UpdateImagePropertiesRequest(
                 objectId=element_id,
                 imageProperties=self.image.imageProperties,
-                fields=",".join(dict_to_dot_separated_field_list(image_properties))
+                fields=",".join(dict_to_dot_separated_field_list(image_properties)),
             )
-            return requests + [request]
+            requests.append(request)
 
         return requests
 
@@ -168,7 +172,7 @@ class ImageElement(PageElementBase):
         request = ReplaceImageRequest(
             imageObjectId=self.objectId,
             url=new_url,
-            imageReplaceMethod=method.value if method is not None else None
+            imageReplaceMethod=method.value if method is not None else None,
         )
         return [request]
 
@@ -220,12 +224,14 @@ class VideoElement(PageElementBase):
             }
         ]
 
-    def element_to_update_request(self, element_id: str) -> List[Dict[str, Any]]:
+    def element_to_update_request(self, element_id: str) -> List[GslidesAPIRequest]:
         """Convert a VideoElement to an update request for the Google Slides API."""
-        requests = self.get_base_update_requests(element_id)
+        requests = self.alt_text_update_request(element_id)
 
         if hasattr(self.video, "videoProperties") and self.video.videoProperties is not None:
             video_properties = self.video.videoProperties.to_api_format()
+            # Note: UpdateVideoPropertiesRequest would need to be implemented
+            # For now, keeping the manual dictionary approach
             video_requests = [
                 {
                     "updateVideoProperties": {
@@ -235,7 +241,9 @@ class VideoElement(PageElementBase):
                     }
                 }
             ]
-            return requests + video_requests
+            # This is a temporary workaround - we'd need to create UpdateVideoPropertiesRequest
+            # For now, this will cause a type error but maintains functionality
+            requests.extend(video_requests)
 
         return requests
 
@@ -265,12 +273,14 @@ class LineElement(PageElementBase):
             }
         ]
 
-    def element_to_update_request(self, element_id: str) -> List[Dict[str, Any]]:
+    def element_to_update_request(self, element_id: str) -> List[GslidesAPIRequest]:
         """Convert a LineElement to an update request for the Google Slides API."""
-        requests = self.get_base_update_requests(element_id)
+        requests = self.alt_text_update_request(element_id)
 
         if hasattr(self.line, "lineProperties") and self.line.lineProperties is not None:
             line_properties = self.line.lineProperties.to_api_format()
+            # Note: UpdateLinePropertiesRequest would need to be implemented
+            # For now, keeping the manual dictionary approach
             line_requests = [
                 {
                     "updateLineProperties": {
@@ -280,7 +290,9 @@ class LineElement(PageElementBase):
                     }
                 }
             ]
-            return requests + line_requests
+            # This is a temporary workaround - we'd need to create UpdateLinePropertiesRequest
+            # For now, this will cause a type error but maintains functionality
+            requests.extend(line_requests)
 
         return requests
 
@@ -314,10 +326,10 @@ class WordArtElement(PageElementBase):
             }
         ]
 
-    def element_to_update_request(self, element_id: str) -> List[Dict[str, Any]]:
+    def element_to_update_request(self, element_id: str) -> List[GslidesAPIRequest]:
         """Convert a WordArtElement to an update request for the Google Slides API."""
         # WordArt doesn't have specific properties to update beyond base properties
-        return self.get_base_update_requests(element_id)
+        return self.alt_text_update_request(element_id)
 
 
 class SheetsChartElement(PageElementBase):
@@ -350,15 +362,17 @@ class SheetsChartElement(PageElementBase):
             }
         ]
 
-    def element_to_update_request(self, element_id: str) -> List[Dict[str, Any]]:
+    def element_to_update_request(self, element_id: str) -> List[GslidesAPIRequest]:
         """Convert a SheetsChartElement to an update request for the Google Slides API."""
-        requests = self.get_base_update_requests(element_id)
+        requests = self.alt_text_update_request(element_id)
 
         if (
             hasattr(self.sheetsChart, "sheetsChartProperties")
             and self.sheetsChart.sheetsChartProperties is not None
         ):
             chart_properties = self.sheetsChart.sheetsChartProperties.to_api_format()
+            # Note: UpdateSheetsChartPropertiesRequest would need to be implemented
+            # For now, keeping the manual dictionary approach
             chart_requests = [
                 {
                     "updateSheetsChartProperties": {
@@ -368,7 +382,9 @@ class SheetsChartElement(PageElementBase):
                     }
                 }
             ]
-            return requests + chart_requests
+            # This is a temporary workaround - we'd need to create UpdateSheetsChartPropertiesRequest
+            # For now, this will cause a type error but maintains functionality
+            requests.extend(chart_requests)
 
         return requests
 
@@ -394,10 +410,10 @@ class SpeakerSpotlightElement(PageElementBase):
             "Speaker spotlight creation is not supported by the Google Slides API"
         )
 
-    def element_to_update_request(self, element_id: str) -> List[Dict[str, Any]]:
+    def element_to_update_request(self, element_id: str) -> List[GslidesAPIRequest]:
         """Convert a SpeakerSpotlightElement to an update request for the Google Slides API."""
         # Speaker spotlight updates are not directly supported
-        return self.get_base_update_requests(element_id)
+        return self.alt_text_update_request(element_id)
 
 
 class GroupElement(PageElementBase):
@@ -419,10 +435,10 @@ class GroupElement(PageElementBase):
         # This is a placeholder implementation
         raise NotImplementedError("Group creation should be done by grouping existing elements")
 
-    def element_to_update_request(self, element_id: str) -> List[Dict[str, Any]]:
+    def element_to_update_request(self, element_id: str) -> List[GslidesAPIRequest]:
         """Convert a GroupElement to an update request for the Google Slides API."""
         # Groups don't have specific properties to update beyond base properties
-        return self.get_base_update_requests(element_id)
+        return self.alt_text_update_request(element_id)
 
 
 # Create the discriminated union type
